fix(maldisplay): fix issue with not being able to increase series progress
If a series had an unknown amount of progress the api reports '0' as the amount of episodes. This just checks if the totalLength is 0, or the progress is less than the totalLength, make sure the +1 icon appears.
This also adds a similar method for decreasing progress, to keep them consistent.