Feat/setup anime view (#121)
* refactor: require passing in userid

Require the user id be passed into the method calls where it is
used instead of relying on it being injected.
This is because the userId is stored inside the database in the app
layer, and we would need to regenerate the bindings when a user logged
in.

* feat: add method to retrieve user id from dao

* feat: use the userid from seriesrepository

Pass in the `UserRepository` and use that to get the `userId` to
use along with the appropriate methods in the `SeriesRepository`.

* fix: dont decrypt empty string

When pulling the preference value, return "" if the string is empty,
instead of trying to decrypt the value.

* feat: setup observer into the anime series

* feat: add adapter for displaying the anime titles

* chore: add glide dependency

* feat: add in network state permissions

Add permission to allow glide to restart failed network requests.

* feat: load the largest image into the view holder

* feat: add button to increment series by 1

* feat: hide the +1 button if reached total length

* refactor: use the poster instead of cover image

* feat: add listener for the adapter

Add a listener the fragment implements so the viewholders can
communicate back on click events.

* chore: delete the assetwizardsettings and ignore

* test: ignore the ui tests for now

Put an @Ignore annotation on the ui tests for now, as currently
they are not setup to run correctly. In the future they should be
setup with some mocks or test harness classes.