feat: hookup the logout button from navigation drawer
Setup the logout button so it can be executed from the Activity. To do this the onClick attribute
had to be added to the menu item, this overrides the behavior of the navigation component which
would not allow a navigation listener to be added (as it only contains a set). This allows it to
quite nicely override the behavior and show the logout button