fix: compilation error due to missing room method
It looks like Room was available from the WorkManager which allowed access to the clearAllTables
method, but with the release of 2.3.0 that method is not accessible. Make the room runtime available
via api so that the app layer can access it. This probably makes more sense anyway, since the app
needs to access room in some way from the database module