feat: add a settings option for the default series state
The option will control what the UserSeriesStatus of a series will be when it is first tracked. The
key for this is stored in core so that other modules can read this value from the preferences, such
as the app-search module. Later on it would make sense to have a new preferences wrapper that sits
in core for the application settings