chore: remove renovate from repository
Renovate looks to be pretty useful as it updated a few dependencies. But the initial PR said it
would update more than it actually updated, and to be honest Dependabot seems to be working fine in
the meantime. Might go back to renovate in the future if dependabot starts to give any issues