refactor: remove user repository from series (#194)
* refactor: remove the user repository from being put into the series repo

The SeriesRepository was being passed the UserRepository, so that it could pull out the userId. This
meant adding a dependency on the Account module where it shouldn't be needed. Add a new Interface
from the Series module for getting the userId from an object. In the app layer add a new package for
binders which will fuse any classes that need to be hooked up to providers

* test: add tests for the UserProviderBinder

Add some basic tests for the binder to ensure its working as intended