test: update code coverage (#306)
* test: update tests to use the global userModel creation method

Instead of mocking the model creation, or having their own methods, use the global method to create
the user model that is stored within the test directory

* test: change usages of SeriesModel to use the global test create

Instead of using mockk or calling their own methods, change usages to all use the global SeriesModel
creation method

* test: increase coverage for the Theme class

Add missing tests in the Theme class

* test: add in tests for the AuthRefreshInterceptor

The interceptor was missing tests, as they were ignored and not written. Take the time to implement
these missing tests

* fix: failure case for refreshing auth crashing

If trying to create a new response to return after failing to refresh the Auth in the
AuthRefreshInterceptor a crash would occur due to not using the api correctly. Fix this to now use
the correct methods as required in the builder

* refactor: remove the need for the MockWebServer in unit tests

The MockWebServer isn't really needed for the unit tests, it should be used more for the integration
tests. Remove it from the Kitsu module for now, and come back to it later for doing integration
testing

* refactor: move ActivityViewModelTests into correct package

File was not in the correct package to align with what it was testing

* chore: add ignore for fragments in code coverage

Fragments will not be covered via unit tests, so ignore them from code coverage. They could be
covered using Espresso tests, but that is not the current plan, so will come back to this at a later
date

* chore: remove App from code coverage

The App.kt class isn't going to be covered by tests, so add it to the ignore for Codecov