refactor: switch the viewmodel creation over to Hilt
Use Hilt to create the ViewModels, this means that the old ViewModelModule and ViewModelFactory can
be removed as they are no longer required. The SavedStateHandle in the ViewModels is not required
either, but from reading it, looks like it takes the value from the Fragment/Activity which could be
handy.