refactor: make first pass at splitting up the api modules more
Currently all of the Kitsu communication exists inside the Kitsu module, if we need to change any of
it it would end up being quite a large job, also injecting it all with Hilt in one go makes the
tests a bit gross. Begin refactoring to split up so there is more modules, one for the API of a
service, and one for the implementation of the service for Kitsu - this means that adding a
different provider later on should be a lot more simple to just plug it in instead of Kitsu.