test: update the sort tests to check the sorted order
Update the SortTest file so that it includes tests that validate the sorted order of the list
afterwards. This is done by having 4 hard coded series objects that are injected into database at
starting time. Using the known data in these entities we can then validate that the expected order
is there after we execute different sorts.