refactor: add search term to results title (#491)
* refactor: change the title of the results page to the search term

Instead of displaying "Results" as the title of the search results, change it to the text the user
input so they know what they were searching for

* fix: search spinner still visible while search text visible

Sometimes the spinner that appears over the search button is still visible when the Search text
becomes visible after successfully searching for series. To negate this, perform the navigation
action first, then hide the spinner, so hopefully the user has already navigated before the spinner
is hidden

* refactor: pass sealed class with event data back to SearchFragment

From the ViewModel pass a sealed class back with the different results that can occur, previously it
was using the AsyncState object and having a success and failure path, but using a single sealed
class allows us to handle them all gracefully. As part of this the search term itself is passed back
as well, this is to stop more inputs on the search bar being added before going to the results page,
which would change the title

* refactor: fix naming of backing variable in viewmodel

* style: fix ktlint issues in the ResultsRobot

* refactor: move the SearchState into its own file and add kdoc