fix: failure to get series if no items
If there was no items in a series list (such as anime), then the call to retrieve the series would
fail. Instead make it return an empty list but still a successful request. This fixes the issue of
if a user has only anime in their library, or only manga.

fix #746