feat: make the new item detail screen and navigate to it
Make the new screen as a full page screen, and add navigation to it. The navigation is done by
adding the nav controller to the module and navigating using that, ideally we would not need to use
this, but it makes it easier and will be removed when it is all swapped to compose correctly.