refactor: use compose for series screens (#928)
* chore: add compose dependencies to series

* feat: add the skeleton files for list use cases

* feat: populate the usercases and make new screen

Make the new compose screen and fragment, and populate the use cases with the required calls.

* feat: emit the list of models from the viewmodel

Do the filtering and sorting in the ViewModel before emitting the ui state.

* feat: update the collection screen to have the list

* feat: add block to check amount of series models

* feat: add pull to refresh to the screen

* feat: show the error snackbar

* fix: viewmodel had wrong hilt annotation

* style: fix a couple of lambda names

* feat: load and map correct seriesType

Load the anime or manga series type depending on how the screen is navigated to.

* feat: update the refreshing status while refreshing series

* refactor: remove the delete series usecase

Remove the usecase, instead delete series will be done on the series detail instead. This will make
it a bit cleaner and the swipe to delete action is a bit weird.

* feat: begin populating the series item

* refactor: create a new model to use with the screen

Make a new data class that will be used to hold the required data for a series in the list.

* feat: set series updating state

* feat: disable the plus one button while updating

* feat: show rating dialog when finishing series

Show the rating dialog if the setting is on and a series is finished.

* feat: begin building the rating dialog screen

* feat: always show rating as what will be sent up

* feat: set string resources for rating dialog

* feat: update the collection view to match old view

* feat: add padding for no series found text

* style: move some code around to make it easier to read

* wip: begin bottom sheet

* chore: change variable name for compose compiler version

* feat: begin adding the sort dialog

Add the menu to the activity that hosts the collection screen, and when its pressed launch the sort dialog.

* refactor: move Composable setup into own files

Move the setup for the sort and rating into their own files to reduce collection screen bloat.

* feat: update the stored sort option when selecting

* feat: sort the series as we update the setting

Switch from the old preferences to using the new datastore so we can get flows. When we update the
sort setting the flow should return to the viewmodel and automatically resort the series.

* refactor: update the rate on completion setting to use datastore

* feat: add the filter to the data store

* refactor: update filter use case to use the datastore method

* feat: prepare all files to show the filter dialog

Prepare the files to show the filter dialog and add any methods that are needed along the way such
as updating the filter.

* feat: update the sort dialog to new style

Update the sort dialog ui, it now more closely matches material design.

* refactor: change the rating dialog buttons to text

* feat: allow updating filters

Setup the filter dialog so it updates the values of the filters once selected. In turn this will
update the preference value which will automatically update the compose screen.

* refactor: move the default UIState setup to the model

* refactor: setup launching the item screen

* feat: make the new item detail screen and navigate to it

Make the new screen as a full page screen, and add navigation to it. The navigation is done by
adding the nav controller to the module and navigating using that, ideally we would not need to use
this, but it makes it easier and will be removed when it is all swapped to compose correctly.

* feat: add a basic item viewmodel

Add basic view model to start building the item screen.

* feat: add error snackbar on the item screen

* feat: build basic item screen

* feat: add menu option to delete series

Add menu option to delete the series instead of swiping away on the list.

* feat: build the delete dialog

Build the delete dialog so it looks like the other dialogs.

* feat: create the layout for the user series status section

* feat: create the layout for the progress

* feat: create the layout for the confirm button

* feat: create the layout for the rating bar

* feat: add loading indicator for the item screen

* fix: strictmode violation due to accessing the dao

* feat: close the screen on successful update

* fix: stop crash if no progress is entered

* refactor: delete the old series detail files

* refactor: delete the old series list items

* chore: remove some now unused dependencies

* refactor: remove unused methods for the SeriesPreferences

* style: fix ktlint issues

* style: fix detekt issues

* test: add tests for the item usecases

* test: add unit tests for the ItemViewModel

* test: add tests for the collection uses cases

* style: fix ktlint issue

* test: temporarily remove the SeriesPreferences tests

* feat: add preview for the FilterDialog

* style: slight code style changes

* fix: issue with missing attribute

* style: code style updates in usecases

* feat: add preview for DeleteDialog

* test: add tests for the collection package

* test: update ui tests for the series screens

* style: fix ktlint issues

* chore: update coverage excludes for compose

* test: fix ui tests running in a block

Reset the series preferences every test to ensure clean state.