refactor: module refactor (#932)
* refactor: move series preferences to new module

Move the SeriesPreferences into a new core:preferences module. This module will be needed to share
any preference between modules and the settings module.

* refactor: move string resources into new module

Move the string resources into a new module only for resource files.

* refactor: remove lots of unused code

Remove code that is not being used. Also remove some of the implementation of the Discover screen
since it is not currently used and is holding a reference to some code that could be removed.

* refactor: move the preference extensions

Move the the SharedPreferenceExtensions file into the new core:preferences module.

* refactor: move the application settings to preference module

Move the ApplicationSettings into the core:preferences module and change its named to
ApplicationPreferences.

* refactor: move theme into flags package

Move the Theme into the flags package which better fits what is is.

* refactor: remove unused chip_text_color

* refactor: remove the now unused AsyncState flag class

* refactor: move the ThemeTests into the correct package

* refactor: move the keys file to preferences module

* refactor: fix namespace for the resources module

* refactor: move the preferences flags to new module

Move the preference flags over to the preferences module where they are used.

* refactor: remove R import

Remove the R import for the preferences since it isn't required.

* style: rename file to fix ktlint

* test: fix the ApplicationPreferences tests

Fix the test having a bad import for the R class.