fix: loading login before series screen
The login screen was being loaded briefly before the series screen, this was because the default
value was to have user not logged in, which is emitted on start. Once the async collection of the
access token was done we could then show the series screen if need be. Changed to wait on a flag for
if it is initialized before loading the initial screen.