refactor: change app module to use compose (#949)
* refactor: remove the url handler from the credentials screen

* refactor: start using hiltViewModel

Update the screens so that they use the new navigation hilt dependency which allows creation of the
view models.

* feat: begin adding new main activity

* feat: load anime or fragment screen based on arg

* feat: add navigation between the series screens

Add the compose navigation between the screens in the series module.

* refactor: tidy up defining the nav graph

* refactor: remove the login fragments

Now the screens can be navigated to with compose, we can remove the fragments.

* feat: load correct screens on start

Load the correct screens when we finish logging in, and initially launching the app.

* fix: incorrect destinations added

* refactor: merge the results into the host screen

Merge the host and results screen into a single screen. This simplifies the navigation and makes
the screen a bit easier to use.

* test: update and fix unit tests

* refactor: remove passing in navController to search

* feat: add a bottom nav to the screen

Add bottom nav to the main activity.

* feat: hide bottom nav on specific screens

Certain screens should hide the bottom navigation bar so they can be full screen.

* feat: add logout option to settings

Add a profile section to the settings screen, which shows the username and avatar image. On this
control also add a logout option.

* feat: add navigation to oss screen

Add navigation to the OSS screen and update the dependency to use the compose version.

* refactor: remove old non-compose files

* chore: remove unused dependencies

* refactor: remove unused hilt module

* refactor: remove classes that are no longer needed

* refactor: remove the not working code to set the theme

Remove the code that sets the theme, this no longer works when using fully compose.

* test: update the LogoutHandler tests

* test: update ui tests to all use compose

* chore: add back in missing accompanist dependency

* chore: remove references to viewBinding

* docs: update documentation on LogoutHandler/Executor

* style: code style update

* refactor: change from using drawables to compose icons

* refactor: move the main activity screen to new file

Split the MainActivity code up into a few different files to keep it tidy.

* style: ktlint fixes

* test: fix the ConfigViewModel tests

* fix: change username input to email

* test: add test for the RetrieveUserUseCase

* feat: handle forced logout

If the token is invalid and we couldnt refresh we need to handle logging the user out so they can
manually log in again.

* fix: loading login before series screen

The login screen was being loaded briefly before the series screen, this was because the default
value was to have user not logged in, which is emitted on start. Once the async collection of the
access token was done we could then show the series screen if need be. Changed to wait on a flag for
if it is initialized before loading the initial screen.

* feat: show the menu buttons on the series list

Show the buttons to sort and filter the series on the collections screen. This allows the old
fragment to be completely removed.

* fix: sort and filter icons wrong

The sort and filter icons were the wrong way around.

* feat: show the delete series option

Show a new button on the series details screen to delete the series, this allows us to remove the
old fragment.

* test: add back in the sort and filter tests

* style: fix ktlint issue