feat: save the links in the db
Save the links into the DB, Moshi was being a pain so pulled gson in for now to get it working which
isnt ideal.
Also made the getSeries call to the repository return a flow, so we can observe a single series in
the detail view and keep getting updates from it.